Prepare for more radical refactoring.
Lots of cruft remains for older versions of CiviCRM. In order to
continue supporting these, we are duplicating our code across v1 and v2
directories. v1 will remain untouched from this point forward, and thus
continue to support older CiviCRMs. v2 will be refactored to better use
Aegir2's coding style, work with the hosting_civicrm front-end module,
and thus allow significant code clean-up.

We are switching which code-base to include based on an option that
will be set by the front-end module. If this variable is unset, or less
than 4.2, the older code will be used.